onvifApp/requirements.txt
django==5.2.5
onvif_zeep==0.2.12
